! fixed a random error exit during the build stage that seems to be caused by environment variables not being set.
> A shell task (/usr/bin/xcrun codesign --force --sign 6EF52167B8457C77203992DC8D52E01193A1E002 --preserve-metadata=identifier,entitlements /Users/qnoid/.windmill/windmill-ios/build/Build/Products/Debug-iphoneos/windmill.app/Frameworks/Alamofire.framework) failed with exit code 72:
xcrun: error: unable to find utility "codesign", not a developer tool or in PATH

The documentation for Process.environment does state that "If this
method isn’t used, the environment is inherited from the process that
created the receiver." which seems to imply that variables are replaced
when set rather than added.

Now passing environment variables as arguments for each process. Will
monitor and report.